On update / delete locations, request submitted for approval

+++ app/Http/Controllers/LocationController.php
@@ -4,13 +4,14 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Location;
+use App\Models\ModificationRequest;
 use Illuminate\Support\Facades\Auth;
 
 class LocationController extends Controller
 {
     public function index()
     {
-        $locations = Location::paginate(10);
+        $locations = Location::with('inProgressReviewRequests')->paginate(10);
 
         return view('locations.index', compact('locations'));
     }
@@ -57,13 +58,27 @@ class LocationController extends Controller
 
         $location = Location::findOrFail($id);
 
-        $location->shop_number = $request->shop_number;
-        $location->floor_number = $request->floor_number;
-        $location->complex = $request->complex;
-        $location->rent = $request->rent;
-        $location->save();
+        if ($location) {
+            if (auth()->user()->hasRole('president')) {
+                $location->shop_number = $request->shop_number;
+                $location->floor_number = $request->floor_number;
+                $location->complex = $request->complex;
+                $location->rent = $request->rent;
+                $location->save();
 
-        return redirect()->route('locations')->with('success', 'Location record updated successfully.');
+                return redirect()->route('locations')->with('success', 'Location record updated successfully.');
+            } else {
+                $this->submitChangeRequest([
+                    "table_name" => 'locations',
+                    "record_id" => $location->id,
+                    "changes" => $request->all(),
+                    "action" => ModificationRequest::REQUEST_TYPE_UPDATE,
+                    "requested_by" => Auth::id(),
+                ]);
+            }
+        }
+
+        return redirect()->route('locations')->with('success', 'Location record updated request submitted successfully.');
     }
 
     public function destroy($id)
@@ -71,13 +86,25 @@ class LocationController extends Controller
         // Find the user by ID
         $location = Location::findOrFail($id);
 
-        // Set the deleted_by field with the authenticated location's ID
-        $location->deleted_by = Auth::id();
-        $location->save(); // Save the $location to update the deleted_by field
+        if ($location) {
+            if (auth()->user()->hasRole('president')) {
+                // Set the deleted_by field with the authenticated location's ID
+                $location->deleted_by = Auth::id();
+                $location->save(); // Save the $location to update the deleted_by field
 
-        // Soft delete the $location
-        $location->delete();
+                // Soft delete the $location
+                $location->delete();
+                return redirect()->route('locations')->with('success', 'Location deleted successfully!');
+            } else {
+                $this->submitChangeRequest([
+                    "table_name" => 'locations',
+                    "record_id" => $location->id,
+                    "action" => ModificationRequest::REQUEST_TYPE_DELETE,
+                    "requested_by" => Auth::id(),
+                ]);
+            }
+        }
 
-        return redirect()->route('locations')->with('success', 'Location deleted successfully!');
+        return redirect()->route('locations')->with('success', 'Location deleted request submitted successfully!');
     }
 }
